Try different import strategy for xfail_without_pywt
